LibSQL using composite domains
