+++ packages/core/src/storage/domains/scores.ts
@@ -1,4 +1,4 @@
-import type { EvalRow } from '../types';
+import type { EvalRow, PaginationArgs, PaginationInfo } from '../types';
 import { MastraStorageBase } from './base';
 import type { MastraStore } from './store';
 
@@ -6,6 +6,11 @@ export abstract class MastraScoresStorage extends MastraStorageBase {
   constructor({ store }: { store: MastraStore }) {
     super({ name: 'SCORES', store });
   }
+
+  abstract getEvals(args: {
+    agentName?: string;
+    type?: 'test' | 'live';
+  } & PaginationArgs): Promise<PaginationInfo & { evals: EvalRow[] }>;
 
   abstract getEvalsByAgentName(agentName: string, type?: 'test' | 'live'): Promise<EvalRow[]>;
 }
